+++ java/javasrc/src/main/java/com/longbridge/SdkNative.java
@@ -238,6 +238,9 @@ public class SdkNative {
 
         public static native void tradeContextCancelOrder(long context, String orderId, AsyncCallback callback);
 
+        public static native void tradeContextCancelOrderAttached(long context, String orderId,
+                        AsyncCallback callback);
+
         public static native void tradeContextAccountBalance(long context, String currency, AsyncCallback callback);
 
         public static native void tradeContextCashFlow(long context, GetCashFlowOptions opts, AsyncCallback callback);